requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
fastapi==0.112.0
fonttools==4.50.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
patsy==0.5.6
pillow==10.2.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyupbit==0.2.33
requests==2.31.0
retrying==1.3.4
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
slack_sdk==3.27.1
sniffio==1.3.1
starlette==0.37.2
statsmodels==0.14.2
threadpoolctl==3.4.0
typing_extensions==4.12.2
tzdata==2024.1
unicorn==2.0.1.post1
urllib3==2.2.1
uvicorn==0.30.5
websockets==12.0
yarl==1.9.4
